Extract middleware, allow use on dynamicRoute
