+++ views/pages/toplevel/funding.njk
@@ -6,7 +6,6 @@
 
 {% block title %}{{ title }} | {% endblock %}
 
-{% set heroImage = heroImages.activePlus %}
 {% set socialImage = heroImage %}
 
 {% block content %}
