+++ views/pages/toplevel/research.njk
@@ -5,7 +5,6 @@
 
 {% block title %}{{ title }} | {% endblock %}
 
-{% set heroImage = heroImages.grassroots %}
 {% set socialImage = heroImage %}
 
 {% block content %}
